+++ src/changelog.md
@@ -1,3 +1,7 @@
+#### 3.1.95
+* added support for labels order (`overrides.chartLegend.legendLabels`) for base charts (#383)
+* added marker option support for charts (#383)
+
 #### 3.1.94
 * fixed issue with `donutChart3D` type